refactor(button-toggle-group): convert to typescript
