+++ src/__internal__/utils/helpers/tags/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./tags";
+export type { TagProps } from "./tags";
