+++ src/components/button-toggle-group/button-toggle-group.style.ts
@@ -2,7 +2,13 @@ import styled, { css } from "styled-components";
 import { StyledButtonToggleLabel } from "../button-toggle/button-toggle.style";
 import ValidationIconStyle from "../../__internal__/validations/validation-icon.style";
 
-const StyledButtonToggleGroup = styled.div`
+import { ButtonToggleGroupProps } from ".";
+import { ValidationProps } from "../../__internal__/validations";
+
+type StyledButtonToggleGroupProps = ValidationProps &
+  Pick<ButtonToggleGroupProps, "inputWidth">;
+
+const StyledButtonToggleGroup = styled.div<StyledButtonToggleGroupProps>`
   display: flex;
 
   ${({ inputWidth }) =>
