--- src/components/button-toggle-group/index.d.ts
@@ -1,1 +1,0 @@
-export { default } from "./button-toggle-group";
