--- src/components/button-toggle-group/index.js
@@ -1,1 +1,0 @@
-export { default } from "./button-toggle-group.component";
